Add first version of all needed core functionality
However, this is not just a raw draft. The main code should be pretty
close to the final result. I might improve the templating flexebility
(maybe adding support for tempalte files instead of using a hardcoded
inline template if no README.md is existing).

Tests will be pushed soon (for sure).