Handle multiple role entry points
This fixes a kind of a leftover from prototyping. DocSmith only processed
the main entry point of a role. This was problematic for a couple of reasons:

a) It incorrectly assumes that defaults/main.yml must exist.
b) Roles can have multiple entry points, not just main.

This commit is a fix for this. DocSmith now parses the argument_spec.yml
and iterate over all defined entry points.

Closes #1.