⚠  Breaking: Rename "BEGIN|END ANSIBLE DOCSMITH" to "BEGIN|END ANSIBLE DOCSMITH MAIN"
As there may be more and more additional content sections in the future, it
would be good to make a small change now - while the tool is still new and has
few users - to avoid bigger breaking changes later.

This is basically a switch to a format like:

<!-- BEGIN ANSIBLE DOCSMITH [Type of content] -->
<!-- END ANSIBLE DOCSMITH [Type of content] -->

What to do?

Replace

   <!-- BEGIN ANSIBLE DOCSMITH -->
   ...
   <!-- END ANSIBLE DOCSMITH -->

With

   <!-- BEGIN ANSIBLE DOCSMITH MAIN -->
   ...
   <!-- END ANSIBLE DOCSMITH MAIN -->

in your documents. That's it. See #8 for more reasoning.